requirements.txt
sentence-transformers
opensearch-py
Pillow
tqdm
exif
jupyterlab
